perf(levm): use signed remaining gas, improving gas cost calculation perf (#4684)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number

<img width="3024" height="3826" alt="image"
src="https://github.com/user-attachments/assets/31ef9703-e365-4c36-821c-df0738237805"
/>

+++ crates/storage/store_db/libmdbx.rs
@@ -1615,7 +1615,7 @@ mod tests {
         Receipt {
             tx_type: TxType::EIP7702,
             succeeded: true,
-            cumulative_gas_used: u64::MAX,
+            cumulative_gas_used: i64::MAX as u64,
             logs,
         }
     }
