+++ tooling/load_test/src/main.rs
@@ -234,7 +234,7 @@ async fn load_test(
                         chain_id: Some(chain_id),
                         value,
                         nonce: Some(nonce + i),
-                        max_fee_per_gas: Some(u64::MAX),
+                        max_fee_per_gas: Some(i64::MAX as u64),
                         max_priority_fee_per_gas: Some(10_u64),
                         gas_limit: Some(TX_GAS_COST * 100),
                         ..Default::default()
